\name{talweg-package}
\alias{talweg-package}
\alias{talweg}
\docType{package}

\title{
	\packageTitle{talweg}
}

\description{
	\packageDescription{talweg}
}

\details{
  To start working with this package, use ?computeForecast.
  Plotting methods are all in 'seealso' section of ?plotError.

  The package R6 is used to abstract data and forecast(er)s implementation from their usage.

	The package devtools could be useful in development stage, since we rely on testthat for
	unit tests and roxygen2 for documentation. rainbow package is used for some plots:
  \code{plotFbox()} and \code{plotFilamentsBox()}.
}

\author{
	\packageAuthor{talweg}

	Maintainer: \packageMaintainer{talweg}
}

%\references{
%	TODO: Literature or other references for background information
%}
